Moved filestore to its own package
