+++ provider/filestore/s3.go
@@ -1,0 +1,76 @@
+package filestore
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	awsv2cfg "github.com/aws/aws-sdk-go-v2/config"
+	"github.com/aws/aws-sdk-go-v2/credentials"
+	s3v2 "github.com/aws/aws-sdk-go-v2/service/s3"
+	"github.com/featureform/provider"
+	"gocloud.dev/blob/s3blob"
+)
+
+type S3FileStoreConfig struct {
+	AWSAccessKeyId string
+	AWSSecretKey   string
+	BucketRegion   string
+	BucketPath     string
+	Path           string
+}
+
+func (s *S3FileStoreConfig) Deserialize(config provider.SerializedConfig) error {
+	err := json.Unmarshal(config, s)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+func (s *S3FileStoreConfig) Serialize() []byte {
+	conf, err := json.Marshal(s)
+	if err != nil {
+		panic(err)
+	}
+	return conf
+}
+
+type S3FileStore struct {
+	Bucket string
+	Path   string
+	GenericFileStore
+}
+
+func (s *S3FileStore) BlobPath(sourceKey string) string {
+	return sourceKey
+}
+
+func NewS3FileStore(config provider.Config) (FileStore, error) {
+	s3StoreConfig := S3FileStoreConfig{}
+	if err := s3StoreConfig.Deserialize(provider.SerializedConfig(config)); err != nil {
+		return nil, fmt.Errorf("could not deserialize s3 store config: %v", err)
+	}
+	cfg, err := awsv2cfg.LoadDefaultConfig(context.TODO(),
+		awsv2cfg.WithCredentialsProvider(credentials.StaticCredentialsProvider{
+			Value: aws.Credentials{
+				AccessKeyID: s3StoreConfig.AWSAccessKeyId, SecretAccessKey: s3StoreConfig.AWSSecretKey,
+			},
+		}))
+	if err != nil {
+		return nil, err
+	}
+	cfg.Region = s3StoreConfig.BucketRegion
+	clientV2 := s3v2.NewFromConfig(cfg)
+	bucket, err := s3blob.OpenBucketV2(context.TODO(), clientV2, s3StoreConfig.BucketPath, nil)
+	if err != nil {
+		return nil, err
+	}
+	return &S3FileStore{
+		Bucket: s3StoreConfig.BucketPath,
+		Path:   s3StoreConfig.Path,
+		GenericFileStore: GenericFileStore{
+			bucket: bucket,
+		},
+	}, nil
+}
